Start GHA workflow to publish to release channel
Creates a new custom workflow that runs after conda-build finishes if it
is on the `main` branch. The idea being this will copy the packages from
`rapidsai-nightly` to `rapidsai`. This does not happen for other
branches than `main` (so `branch-YY.MM` would be excluded and only
publish to the nightly channel). This would help automate the final step
to get XGBoost packages onto the `rapidsai` channel.