Only use `cccl` package on Windows
As Linux builds now pick up CCCL via RMM's patched copy, drop the
unneeded `cccl` package from Linux builds.

Windows still needs this as it the CCCL in the CUDA Toolkit does not
work correctly there. Fortunately the `cccl` package works just fine and
the version is compatibility with the version in the CUDA Toolkit.