Bump `build/number` to `2`
To make it easier to pick out the 25.12 packages from the 26.02
packages, bump the `build/number` to `2`.

This is not strictly needed as the package string hash is updated when
the RMM version changes. However this makes it a bit easier to pick out
which builds belong in the 25.12 release.